Move palette code into a sensible place
